+++ src/api.rs
@@ -17,6 +17,7 @@ use crate::{
         lookup_capabilities, lookup_credits, lookup_id, lookup_pubkey, Capabilities,
         LookupCriterion,
     },
+    receive::IncomingMessage,
     types::{BlobId, FileMessage, MessageType},
     SecretKey, MSGAPI_URL,
 };
@@ -350,6 +351,18 @@ impl E2eApi {
         )
         .await
     }
+
+    /// Deserialize an incoming Threema Gateway message in
+    /// `application/x-www-form-urlencoded` format.
+    ///
+    /// This will validate the MAC. If the MAC is invalid,
+    /// [`ApiError::InvalidMac`] will be returned.
+    pub fn decode_incoming_message(
+        &self,
+        bytes: impl AsRef<[u8]>,
+    ) -> Result<IncomingMessage, ApiError> {
+        IncomingMessage::from_urlencoded_bytes(bytes, &self.secret)
+    }
 }
 
 /// A convenient way to set up the API object.
@@ -362,9 +375,9 @@ impl E2eApi {
 /// use threema_gateway::{ApiBuilder, SimpleApi};
 ///
 /// let gateway_id = "*3MAGWID";
-/// let gateway_secret = "hihghrg98h00ghrg";
+/// let api_secret = "hihghrg98h00ghrg";
 ///
-/// let api: SimpleApi = ApiBuilder::new(gateway_id, gateway_secret).into_simple();
+/// let api: SimpleApi = ApiBuilder::new(gateway_id, api_secret).into_simple();
 /// ```
 ///
 /// ## E2E API
@@ -373,10 +386,10 @@ impl E2eApi {
 /// use threema_gateway::{ApiBuilder, E2eApi};
 ///
 /// let gateway_id = "*3MAGWID";
-/// let gateway_secret = "hihghrg98h00ghrg";
+/// let api_secret = "hihghrg98h00ghrg";
 /// let private_key = "998730fbcac1c57dbb181139de41d12835b3fae6af6acdf6ce91670262e88453";
 ///
-/// let api: E2eApi = ApiBuilder::new(gateway_id, gateway_secret)
+/// let api: E2eApi = ApiBuilder::new(gateway_id, api_secret)
 ///                              .with_private_key_str(private_key)
 ///                              .and_then(|builder| builder.into_e2e())
 ///                              .unwrap();
@@ -460,6 +473,8 @@ impl ApiBuilder {
     }
 
     /// Return a [`E2eAPI`](struct.SimpleApi.html) instance.
+    ///
+    /// This will fail if no private key was set.
     pub fn into_e2e(self) -> Result<E2eApi, ApiBuilderError> {
         match self.private_key {
             Some(key) => Ok(E2eApi::new(